requirements.txt
Flask
Flask-WTF
Flask-Login
Flask-SQLAlchemy
Werkzeug
WTForms
email-validator
python-dotenv
psycopg2-binary